rename: 도메인 하위 폴더로 이동
