+++ src/main/java/com/Wavey/WaveyService/domain/route/dto/RouteResponse.java
@@ -1,4 +1,4 @@
-package com.Wavey.WaveyService.route.dto;
+package com.Wavey.WaveyService.domain.route.dto;
 
 import lombok.*;
 import java.time.LocalDateTime;
