+++ src/main/java/com/Wavey/WaveyService/domain/route/repository/RouteRepository.java
@@ -1,6 +1,6 @@
-package com.Wavey.WaveyService.route.repository;
+package com.Wavey.WaveyService.domain.route.repository;
 
-import com.Wavey.WaveyService.route.entity.Route;
+import com.Wavey.WaveyService.domain.route.entity.Route;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
